+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* **Breaking** Refactor `MergeVarBuilder` to avoid cloning all input values.
 * **Breaking** Remove `zng::app::crash_handler::{BacktraceFrame, CrashPanic}`.
     - Unified with `zng::task::process::tap::PanicInfo` and related types.
 * **Breaking** Remove deprecated `ProcessStartArgs::yield_until_app`.